Move Archive.php archive invalidation to Loader so we only invalidate when about to launch archiving.
